+++ src/cr0s/WarpDrive/SpaceProvider.java
@@ -2,6 +2,7 @@ package cr0s.WarpDrive;
 
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
+import net.minecraft.block.Block;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.util.ChunkCoordinates;
@@ -109,7 +110,7 @@ public class SpaceProvider extends WorldProvider
 
         for (int var2 = 0; var2 <= 15; ++var2)
         {
-            float var3 = 1.0F - var2 / 15.0F;
+            float var3 = 1.0F - (float)var2 / 15.0F;
             this.lightBrightnessTable[var2] = (1.0F - var3) / (var3 * 3.0F + 1.0F) * (1.0F - var1) + var1;
         }
     }
@@ -132,14 +133,14 @@ public class SpaceProvider extends WorldProvider
     {
         setCloudRenderer(new CloudRenderBlank());
         //setSkyRenderer(new SpaceSkyRenderer());
-        return this.worldObj.getWorldVec3Pool().getVecFromPool(0.0D, 0.0D, 0.0D);
+        return this.worldObj.getWorldVec3Pool().getVecFromPool((double) 0, (double) 0, (double) 0);
     }
 
     @SideOnly(Side.CLIENT)
     @Override
     public Vec3 getFogColor(float par1, float par2)
     {
-        return this.worldObj.getWorldVec3Pool().getVecFromPool(0.0D, 0.0D, 0.0D);
+        return this.worldObj.getWorldVec3Pool().getVecFromPool((double) 0, (double) 0, (double) 0);
     }
 
     @SideOnly(Side.CLIENT)
